use usize

+++ crates/engine/primitives/src/config.rs
@@ -9,8 +9,10 @@ pub const DEFAULT_MEMORY_BLOCK_BUFFER_TARGET: u64 = 2;
 /// Default maximum concurrency for proof tasks
 pub const DEFAULT_MAX_PROOF_TASK_CONCURRENCY: u64 = 256;
 
-/// Default number of reserved CPU cores for non-reth processes
-pub const DEFAULT_RESERVED_CPU_CORES: u64 = 1;
+/// Default number of reserved CPU cores for non-reth processes.
+///
+/// This will be deducated from the thread count of main reth global threadpool.
+pub const DEFAULT_RESERVED_CPU_CORES: usize = 1;
 
 const DEFAULT_BLOCK_BUFFER_LIMIT: u32 = 256;
 const DEFAULT_MAX_INVALID_HEADER_CACHE_LENGTH: u32 = 256;
@@ -70,7 +72,7 @@ pub struct TreeConfig {
     /// Maximum number of concurrent proof tasks
     max_proof_task_concurrency: u64,
     /// Number of reserved CPU cores for non-reth processes
-    reserved_cpu_cores: u64,
+    reserved_cpu_cores: usize,
 }
 
 impl Default for TreeConfig {
@@ -107,7 +109,7 @@ impl TreeConfig {
         cross_block_cache_size: u64,
         has_enough_parallelism: bool,
         max_proof_task_concurrency: u64,
-        reserved_cpu_cores: u64,
+        reserved_cpu_cores: usize,
     ) -> Self {
         Self {
             persistence_threshold,
@@ -156,7 +158,7 @@ impl TreeConfig {
     }
 
     /// Return the number of reserved CPU cores for non-reth processes
-    pub const fn reserved_cpu_cores(&self) -> u64 {
+    pub const fn reserved_cpu_cores(&self) -> usize {
         self.reserved_cpu_cores
     }
 
@@ -265,7 +267,7 @@ impl TreeConfig {
     }
 
     /// Setter for the number of reserved CPU cores for any non-reth processes
-    pub const fn with_reserved_cpu_cores(mut self, reserved_cpu_cores: u64) -> Self {
+    pub const fn with_reserved_cpu_cores(mut self, reserved_cpu_cores: usize) -> Self {
         self.reserved_cpu_cores = reserved_cpu_cores;
         self
     }
